requirements.txt
flet==0.24.1
pandas>=2.1
pyserial>=3.5
